refactors server modules for bundling with webpack
