+++ package.json
@@ -6,10 +6,12 @@
   "proxy": "http://localhost:3004",
   "scripts": {
     "start": "npm run build && npm run start-server",
-    "start-server": "node ./build/server/index.js",
+    "start-server": "node ./build/server/bundle.js",
     "build": "npm run build-server && npm run build-client",
     "build-client": "react-scripts build",
-    "build-server": "tsc -p ./tsconfig.server.json",
+    "build-server": "npm run ts-compile-server && npm run webpack-server",
+    "webpack-server": "node ./build/server/server/pack.js",
+    "ts-compile-server": "tsc -p ./tsconfig.server.json",
     "dev": "concurrently --names 'FRONT,BACK' --prefix-colors 'blue,yellow' \"npm run dev-client\" \"npm run dev-server\"",
     "dev-client": "react-scripts start",
     "dev-server": "nodemon ./src/server/index.ts",
@@ -54,6 +56,7 @@
     "@types/react-router-dom": "^5.3.3",
     "@types/uuid": "^8.3.4",
     "@types/web-push": "^3.3.2",
+    "@types/webpack-node-externals": "^3.0.0",
     "concurrently": "^6.2.0",
     "marked": "^4.0.12",
     "nodemon": "^2.0.20",
@@ -64,7 +67,8 @@
     "react-scripts": "^5.0.1",
     "sass": "^1.34.0",
     "ts-node": "^10.4.0",
-    "typescript": "^4.5.5"
+    "typescript": "^4.5.5",
+    "webpack-node-externals": "^3.0.0"
   },
   "eslintConfig": {
     "extends": [
