+++ tsconfig.server.json
@@ -15,11 +15,9 @@
       "client/*": ["./client/*"],
       "server": ["./server"],
       "server/*": ["./server/*"]
-    }
+    },
+    "jsx": "react-jsx"
   },
-  "include": ["src/server"],
-  "exclude": [
-    "*.test.ts",
-    "**/*.test.ts",
-  ]
+  "include": ["src"],
+  "exclude": ["*.test.ts", "**/*.test.ts"]
 }
